+++ model/scikit/examples/lr/test_data.sh
@@ -1,0 +1,5 @@
+cat > test.csv << EOF
+Years,Expertise,Trust,Salary
+4,9,0.5,50
+5,11,0.6,60
+EOF